Completed part1 and user register logic

+++ src/utils/cloudinary.js
@@ -4,9 +4,9 @@ import fs from "fs"
 
 // Configuration
 cloudinary.config({ 
-    cloud_name: 'process.env.CLOUDINARY_CLOUD_NAME', 
-    api_key: 'process.env.CLOUDINARY_API_KEY', 
-    api_secret: 'process.env.CLOUDINARY_API_SECRET' 
+    cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
+    api_key: process.env.CLOUDINARY_API_KEY, 
+    api_secret: process.env.CLOUDINARY_API_SECRET 
 });
     
 
@@ -18,11 +18,15 @@ const uploadOnCloudinary = async (localFilePath) => {
             resource_type: "auto",
         })
         //file has been uploaded successfully 
-        console.log("File uploaded successfully on cloudiinary : ", response.url);
+        // console.log("File uploaded successfully on cloudiinary : ", response.url);
+        fs.unlinkSync(localFilePath);
         return response;
     }
     catch(error){
-        fs.unlinkSync(localFilePath); // Delete the file from local storage as uploading on cloudinary failed -- did it in Synchronous way i.e yeah hona hi chahiye tabhi aage badhenege 
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        }
+        console.error("Upload failed:", error);
         return null;
     }
 }
